+++ rtl/ibus/sys_conf.vhd
@@ -1,6 +1,6 @@
--- $Id: sys_conf.vhd 1043 2018-09-09 10:20:12Z mueller $
+-- $Id: sys_conf.vhd 1111 2019-02-10 16:13:55Z mueller $
 --
--- Copyright 2015-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+-- Copyright 2015-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 --
 -- This program is free software; you may redistribute and/or modify it under
 -- the terms of the GNU General Public License as published by the Free
@@ -16,9 +16,10 @@
 -- Description:    Default definitions for ibdr_maxisys
 --
 -- Dependencies:   -
--- Tool versions:  xst 14.7; viv 2014.4-2016.4; ghdl 0.18-0.33
+-- Tool versions:  xst 14.7; viv 2014.4-2018.3; ghdl 0.18-0.35
 -- Revision History: 
 -- Date         Rev Version  Comment
+-- 2019-02-09  1110   1.1    use typ for DL,PC,LP; add dz11
 -- 2018-09-08  1043   1.0.2  add sys_conf_ibd_kw11p
 -- 2017-01-29   847   1.0.1  add sys_conf_ibd_deuna
 -- 2015-03-14   658   1.0    Initial version 
@@ -32,9 +33,12 @@ use work.slvtypes.all;
 package sys_conf is
 
   -- configure character and communication devices
-  constant sys_conf_ibd_dl11_1 : boolean := true;  -- 2nd DL11
-  constant sys_conf_ibd_pc11   : boolean := true;  -- PC11
-  constant sys_conf_ibd_lp11   : boolean := true;  -- LP11
+  -- typ for DL,DZ,PC,LP: -1->none; 0->unbuffered; 4-7 buffered (typ=AWIDTH)
+  constant sys_conf_ibd_dl11_0 : integer :=  4;    -- 1st DL11
+  constant sys_conf_ibd_dl11_1 : integer :=  4;    -- 2nd DL11
+  constant sys_conf_ibd_dz11   : integer :=  4;    -- DZ11
+  constant sys_conf_ibd_pc11   : integer :=  4;    -- PC11
+  constant sys_conf_ibd_lp11   : integer :=  5;    -- LP11
   constant sys_conf_ibd_deuna  : boolean := true;  -- DEUNA
 
   -- configure mass storage devices